feat: create StockMutation logic

+++ src/main/java/com/dak/spravel/dto/request/order/OrderItemsRequest.java
@@ -2,12 +2,10 @@ package com.dak.spravel.dto.request.order;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class OrderItemsRequest {
     private Long orderId;
     private Long productId;
 
-    private BigDecimal qty;
+    private Long qty;
 }
